requirements.txt
geographiclib==2.0
geopy==2.4.1
pyswisseph==2.10.3.2
pytz==2025.2
